refactoring classes endpoint

+++ src/main/java/com/jj/Gradebook/service/classes/ClassServiceImpl.java
@@ -2,6 +2,8 @@ package com.jj.Gradebook.service.classes;
 
 import com.jj.Gradebook.dao.ClassRepository;
 import com.jj.Gradebook.entity.Class;
+import com.jj.Gradebook.exceptions.EntityAlreadyExistException;
+import com.jj.Gradebook.exceptions.EntityNotFoundException;
 import jakarta.transaction.Transactional;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -17,32 +19,47 @@ public class ClassServiceImpl implements ClassService{
 
     @Override
     public List<Class> findAll() {
-        return classRepository.findAll();
+        List<Class> data = classRepository.findAll();
+
+        return data;
     }
 
     @Override
-    public Class findById(int id) {
+    public Class findById(int id) throws EntityNotFoundException {
         Optional<Class> result = classRepository.findById(id);
 
         Class theClass;
         if(result.isPresent()){
             theClass = result.get();
         }
-        else { //TODO: FIND BETTER APPROACH
-            throw new RuntimeException("No class with id - " + id);
+        else {
+            throw new EntityNotFoundException("No class with id - " + id);
         }
         return theClass;
     }
 
     @Override
     @Transactional
-    public Class save(Class classes) {
-        return classRepository.save(classes);
+    public Class save(Class classes) throws EntityAlreadyExistException {
+        Optional<Class> existingClass = classRepository.findClassByClassNameAndStartYear(classes.getClassName(),classes.getStartYear());
+
+        if (existingClass.isEmpty()){
+            return classRepository.save(classes);
+        }
+        else {
+            throw new EntityAlreadyExistException("Class already exists!");
+        }
     }
 
     @Override
     @Transactional
-    public void deleteById(int id) {
-        classRepository.deleteById(id);
+    public void deleteById(int id) throws EntityNotFoundException {
+        Optional<Class> existingClass = classRepository.findById(id);
+        if (existingClass.isPresent()){
+            classRepository.deleteById(id);
+        }
+        else {
+            throw new EntityNotFoundException("No class with id - " + id);
+        }
     }
 }
